+++ libraries/treeLayout/src/Layout.cpp
@@ -30,17 +30,17 @@ namespace treeLayout
         _offset = value;
     }
 
-    Layout::Layout(uint64_t size, double min_offset, double max_offset, double min_depth, double max_depth)
+    Layout::Layout(size_t size, double min_offset, double max_offset, double min_depth, double max_depth)
         : _positions(size), _minOffset(min_offset), _maxOffset(max_offset), _minDepth(min_depth), _maxDepth(max_depth)
     {
     }
 
-    VertexPosition& Layout::operator[](uint64_t index)
+    VertexPosition& Layout::operator[](size_t index)
     {
         return _positions[index];
     }
 
-    const VertexPosition& Layout::operator[](uint64_t index) const
+    const VertexPosition& Layout::operator[](size_t index) const
     {
         return _positions[index];
     }
